backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
azure-storage-blob==12.19.0
pillow==10.1.0
